+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
 - Change `TimeKind::default` to `TimeKind::Real`.
 - Add `TimeKind` configuration to `SmoothNudge`.
 - Add `TimeKind` configuration to `DeltaScale`.