+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.21.0] - 2025-12-17
+
+### Changed
+
+- Update to Bevy 0.18.0.
+
 ## [0.20.1] - 2025-12-29
 
 ### Changed
@@ -69,8 +75,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `Cancelled` -> `Cancel`.
   - `Completed` -> `Complete`.
   - `Started` -> `Start`.
-  Note that `Press`, `Release`, and `Cancel` collide with names from `bevy_picking` and present in both `bevy::prelude::*` and `bevy_enhanced_input::prelude::*`.
-  To disambiguate, import `bevy_enhanced_input::prelude::{*, Press, Release, Cancel}`.
+    Note that `Press`, `Release`, and `Cancel` collide with names from `bevy_picking` and present in both `bevy::prelude::*` and `bevy_enhanced_input::prelude::*`.
+    To disambiguate, import `bevy_enhanced_input::prelude::{*, Press, Release, Cancel}`.
 - Serde integration is now gated behind the `serialize` feature.
 - Warn on dimension mismatch instead of panicking.
 
@@ -493,7 +499,8 @@ This update features a big rewrite into a component-based API. The core concepts
 
 Initial release.
 
-[unreleased]: https://github.com/simgine/bevy_replicon/compare/v0.20.1...HEAD
+[unreleased]: https://github.com/simgine/bevy_replicon/compare/v0.21.0...HEAD
+[0.21.0]: https://github.com/simgine/bevy_replicon/compare/v0.20.1...v0.21.0
 [0.20.1]: https://github.com/simgine/bevy_replicon/compare/v0.20.0...v0.20.1
 [0.20.0]: https://github.com/simgine/bevy_replicon/compare/v0.19.3...v0.20.0
 [0.19.3]: https://github.com/simgine/bevy_replicon/compare/v0.19.2...v0.19.3